Differences in classlib shutdown
In JNI DestroyJavaVM(), after waiting for non-daemon threads to
die we need to run registered shutdown handlers.

With GNU Classpath, the only way to run the handlers is to call
System.exit().  This, however, calls back into the VM to exit
the process (i.e. we never return back from DestroyJavaVM).

On OpenJDK, the package-private java.lang.Shutdown::shutdown()
can be ran to only run the shutdown handlers.  This means we
return from DestroyJavaVM back to the java wrapper, thus enabling
any error messages to be shown.  This class is JDK specific, and
is not part of the Java spec.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>