Fix bug in JVM_NewMultiArray
JVM_NewMultiArray was broken for non-primitive classes, as the
array name for Object array classes was not constructed correctly.
Also forgot to check for a negative dimension in the dimensions
array, and that the classloader may be other than the bootstrap
classloader.  I implemented the function when I was tired one night
and it shows!

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>