Rewrite integer divide overflow check
Previous check-in was incomplete, as it didn't compile with
stack-caching enabled.  It also didn't work with stack-caching
enabled and overflow checking on (in one case, the macro
is passed an argument with pointer update, and when checking
is enabled the argument is evaluated multiple times).  I also
didn't like the messy "is_div" parameter.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>