Rework OpenJDK storage of native thread structure
In many instances we need to be able to map from a Java level
thread object to the underlying native thread structure.  The
obvious way to do this is to store the native pointer in the
thread object in some private "VM data" field.  However, when
first implementing this, there was no obvious field within the
thread object to do this.  However, the VM maintains a hash
table mapping thread IDs to native thread pointers, and so this
was used to do the mapping from thread object to Thread*.

This, however, is inefficient as it requires retrieving the
thread ID, looking up the thread ID in the table, locking the
table, etc. (luckily it's not used in critical code).

There obviously must be a better way.  From examining HotSpot,
it transpires that it uses the "eetop" field for the mapping.
This change makes JamVM use the eetop field as well.  This change
also aligns the OpenJDK code much closer to the GNU Classpath code,
which stores the Thread* in the VMThread object, which is referenced
from the thread object.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>