Fix leak of native thread structure
When a Java thread dies, we need to release the native thread
structure.  However, we need to be careful, as another thread may
be concurrently accessing it.  The solution taken in JamVM is
to free the structure when the Java thread is GCed.  However,
without an explicit link between the Java thread and the native
structure this couldn't be done, and the native thread structure
was leaked.  This has been fixed, so we can free the thread
structures.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>