Unify command line options parsing
JamVM has two sets of options parsing.  The first is in the jamvm
executable that is the VM entry-point when used with GNU Classpath
(with GNU Classpath, JamVM is linked into a single standalone
executable, as well as libjvm.so).

The other is the options parsing within the JNI invocation API.
This entry-point is used when the VM is linked into another
process, and the VM is created via JNI.  This is the standard
entry point when used with OpenJDK/IcedTea.  Here, the VM
is started via a small launcher executable, which handles command
line parsing, passing arguments to the JNI invocation API.

Within these two routines, many of the options can be handled
identically.  However, small differences has made separating these
options out into a common routine difficult.  These issues have
now been resolved and a common routine for parsing is now used.

The motivation for the change is the OpenJDK port.  In the past,
the JNI invocation API only handled a subset of the special -X
options, but with the OpenJDK port, this is now the main entry-
point into the VM.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>