Fix backwards cache conflict resolution code
Direct-threaded interpretation with stack-caching requires the
cache depth to be determined statically at each bytecode.  This
is done via abstract interpretation of the bytecodes at re-
writing.

With control-flow (branches, etc.) a bytecode may be reached via
multiple paths. If the cache depth is different between the paths
we have a conflict which must be resolved.  Conflicts may occur
in a backwards or forwards direction depending on the branch.

The resolution of backwards conflicts contained a "one-off" bug.
This has existed since the code was written in 2005.  It has
not been seen as javac does not produce code which causes a
conflict.  However, the SPECjvm 2008 xml.transform benchmark
auto-generates bytecode.  This was triggering a conflict on
a backwards GOTO (and this only occurs once).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>