Fix callJNIMethod on i386 with -fomit-frame-pointer
The callJNIMethod on i386 used to rely on the use of a
frame-pointer to refer to arguments (it modified the sp
invisibly to gcc).

I have been aware of the problem for several years as
JamVM did not work (on i386) when -fomit-frame-pointer
was used.  It has now become a problem (prompting a
fix) as gcc 4.6 has -fomit-frame-pointer on by default.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>