Initial implementation of JVM_FindClassFromBootLoader
OpenJDK 7 has an additional interface function (due to
problems with visibility of JVM_FindClassFromClassLoader).

In the past JVM_FindClassFromBootLoader used the same
logic as JVM_FindClassFromClassLoader, but with a null
loader.  Recent changes in hotspot, however, has
modified it to not throw an exception but simply to
return NULL if the class doesn't exist (so the throwError
param disappears).

Unfortunately, JamVM's class-loader throws an exception
and returns NULL.  So for the present we clear any
thrown exception.  However, this should be fixed as it's
inefficient.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>