Enable shutdownVM to be called with OpenJDK classlib
Minor refactor of shutdown code to allow shutdownVM to be called
when using OpenJDK class-library.  Previously, shutdownVM exited.
With OpenJDK, shutdown is done via JNI DestroyJavaVM.  This needs
to return back to the launcher and so can't exit.

As shutdownVM is now called, details of how much code memory was
used can now be obtained via -Xcodestats.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>