Make classlib init functions consistent + warnings
Many module initialisation functions (e.g. initialiseFoo)
call a classlib dependent initialisation function (classlibInitialiseFoo).
When doing the re-factoring, I alternated between having the
functions return an error condition, or simply handling the error
themselves (terminating the VM).

This change makes them consistently return an error condition.
Currently, the upper initialiseFoo() just exits, but eventually
the error condition should be propogated all the way up to the caller
of initVM().  For OpenJDK, this means JNI_CreateJavaVM() will return
an error which will be handled by the launcher.  In reality,
it's not a huge issue, as the launcher simply prints "unable to create
VM" and exits itself.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>