Propogate initialisation errors to top-level
Previously, initialisation errors encountered while initialising
the individual VM modules caused the VM to exit within the module.
This was done as there was no value in propogating the error to
the top-level (as it would only exit then).  However, with OpenJDK,
the VM is created via JNI.  This can return an error to the calling
program (in this case the launcher).  This will only abort, but
propogating the error is cleaner than aborting within the VM.