Fix class GC with classes containing Miranda methods
Miranda methods are created by the VM for classes with unimplemented
interface methods.  These methods are shallow copies of the interface
method, and contain shared throw tables.  Therefore, when freeing
a Miranda method, the shared data must not be freed (it will be freed
if and when the interface class is freed).

This fixes invalid frees seen while building packages such as jaxme
(the data was being freed when freeing the Miranda method, and
then freed again when freeing the interface, leading to a double
free).