Intern strings when creating a StackTraceElement
A StackTraceElement contains strings for file name, class name
and method name.  When creating a StackTraceElement, strings
were created for these but they weren't interned.  That is,
two StackTraceElements could have the same file name as they
have the same sequence of characters (String.equals), but the
instances may not be the same (ptr1 == ptr2).

As far as I can see, this is within the Java spec which states
equality for elements as:

     static boolean equals(Object a, Object b) {
         return a==b || (a != null && a.equals(b));
     }

However, a jtreg test is failing, as it is comparing a method name
to a string via ste.getMethodName() == "xxx".  Of course, the
"reference implementation" interns the strings so this will work.
However, it is a bad test, as it relies on RI behaviour and not
the Java spec.

I have two choices.  Get the test fixed, or intern the strings.
As other programs may rely on RI behaviour, I have decided to
intern the strings.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>