Fix typo in definition of ACC_MIRANDA
The VM creates methods internally for unimplemented
interface methods (so called Miranda methods).  These
are marked via a bit in the method access flags.

This bit should have been an unused bit in the access flags
(0x8000).  But due to a typo (a long time ago) it was defined
as 0x0800.  This conflicts with ACC_STRICT (strictfp method).

Miranda methods are not returned in getDeclaredMethods().
The typo means strictfp methods look like Miranda methods
and aren't returned.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>