Tidy up JNI stubs mechanism
In 2010 I introduced a JNI stubs mechanism that enables stubs to
be created for common JNI method signatures.  This has never been
used (the tables are empty), as it was difficult to tell which stubs
were most used.

The stubs generator has now been extended to enable the stubs to be
profiled.  Also, instead of the stubs having to be machine-
dependent, generic stubs can now be used with or without FFI (generic
stubs are pretty good for most platforms, and enables them to be
used without having to write support for each platform).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>