Skip Java-reflection-related DelegatingClassLoaders
Once upon a time, the reflection support in Hotspot was done
natively.  Then somebody had the bright idea of doing it in
Java via proxies created on the fly (auto-generated bytecodes).

Nice idea in principle but it's not invisible to the VM.  This
needs to know about the Java proxy stuff in order to handle it
specially in places such as stack-walking, access control, etc.

One place even Sun forgot about.  In order to not pollute the
target classloaders, the proxy stuff got their own classloader
which simply delegates (DelegatingClassLoader).  On the surface,
this should be invisible to the VM but it's not (see Sun Bug
4474172).

The upshot is that JamVM needs to skip DelegatingClassLoaders
during class-loading, and use the parent loader instead.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>