Add missing REF_TO_OBJs for references passed from JNI
In JamVM, JNI references are real direct object references
and not handles.  However, the bottom bits are tagged to
record the reference type (Global, Weak, etc.).  This
means the bits must be masked off whenever a reference
is passed from a JNI method and is used within the VM.

Somehow, I forgot to mask the bits off when a reference
is returned from a JNI method, and also when a reference
is stored into an object field or an array.  This has
been present since the start of 2010, so I'm surprised
it has taken so long to show up!

The error showed up with Qt Jambi, which is a wrapper
which enables Qt to be used from Java (crash reported
by Xerxes Ranby).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>