Generate machine-dependent stubs for i386.
Because all parameters are passed on the stack,
a much smaller number of stubs are necessary
on i386 to cover the same set of signatures as
the "generic" JNI stubs.  For this reason,
i386 is the only platform that has machine-
dependent stubs (all platforms can support
machine dependent stubs, but other platforms
offer no advantage over the generic stubs).

However, although machine-dependent stubs are
on by default for i386, I never bothered to
generate them (it previously had empty stub
tables)!  This was mainly because I believed
nobody really still used i386 instead of
amd64.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>