When suspending, ignore detached threads that have died
When signalling threads to suspend, if pthread_kill returns
ESRCH this indicates that the thread has died.  This can only
occur when an external thread has been attached to the VM
via JNI and it has exited without detaching.  Although it
is a user error, it will deadlock the suspension code as it
will hang waiting for the thread to suspend.  Set the state
to BLOCKING, to ignore the thread. Note, no attempt is made
to clean-up the error; the thread will still appear to be
"live" (as with Hotspot).  We simply stop the thread from
hanging the VM.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>