Changes for Raspbian (Raspberry Pi)
Debian, when producing a new high-performance distribution
for modern ARM cores, decided to introduce a new "architecture"
called armhf (as opposed to armel).  This supports VFP for
floating-point, and in addition uses floating-point registers
for passing floating-point arguments (hence hf for hard-float).

For armhf, they set a minimum requirement of ARMv7-A and VFPv3-D16.
Unfortunately, the Raspberry Pi does not meet this (being ARMv6 +
VFPv2).  So Raspbian is a recompile of Debian Wheezy armhf for
ARMv6 + VFPv2.

Unfortunately, when I ported JamVM to hard-float ABI, I decided
to use Thumb2.  This was fine, as armhf must support Thumb2.
However, this decision means JamVM doesn't work when compiled
for Raspbian as it contains ARMv7-A instructions.  To make it
worse, I initially implemented hard-float ABI in ARM mode, but
decided to switch to Thumb2 as it was a good opportunity to use
the tbb instruction!

As the Thumb2 implementation will have better performance on ARMv7-A
cores than the ARM version, I have decided to also include the
ARM version which will be used when ARMv7-A isn't supported (i.e.
Raspbian).

In addition, I have updated the code to use ldrex/strex for atomic
operations instead of swp on ARMv6.  Previously, ldrex/strex was
only used on ARMv7-A, but swp is also deprecated on ARMv6.  I
previously only did it for ARMv7-A because of concerns that some
ARMv6 implementations didn't implement ldrex/strex properly
(e.g. OMAP2).  I can no longer find any evidence for this.  Also,
OMAP2 is now pretty ancient, and I prefer to fix it for the buggered
implementations, if and when a problem occurs, rather than nobbling
it for all ARMv6.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>