Do not put version number in libjvm.so SONAME
In OpenJDK the libjava.so shared object has a dependency on libjvm.so.
Hotspot's libjvm.so SONAME doesn't have a version number but JamVMs
does (created via libtool).  If JamVM is being built as an additional
VM, libjava.so is linked against Hotspot's libjvm.so, and we end up with
a library which can't find JamVMs libjvm.so.  However, it works if
JamVM is the only VM being built, as libjava.so is linked against
JamVMs libjvm.so.

Simple fix is to stop libtool versioning the SONAME with -avoid-version.
Patch provided by Xerxes Ranby and Matthias Klose.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>