Dummy implementation of sun.misc.Perf natives
The OpenJDK 7 and OpenJDK 8 sources contain various calls
to sun.misc.Perf for performance profiling of class-loading.
Whether this is left-over debug code or is intentional is
not clear.  However, this causes JamVM to fail as it doesn't
implement the native methods for sun.misc.Perf.

IcedTea 7 contains a patch to remove these calls to enable
JamVM to run.  But this doesn't help when running with
"vanilla" OpenJDK 8, which I'm using for the JSR 292 work
(no IcedTea version for OpenJDK 8 exists, that I'm aware off).

Hacking the sources to remove these calls everytime I get
a new version is a huge pain-in-the-ass.  So this is a dummy
implementation, sufficient to get the sources to run without
hacking them.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>