Very simple field injection framework
Currently there's only one field that needs to be injected
into one class (java.lang.invoke.MemberName).  Plus of
course it's not needed for GNU Classpath (as the Java classes
are part of JamVM).  For this reason I have made the
"framework" as light as possible, and it will be entirely
optimised out for GNU Classpath.  However, it is a proper
framework, and extra classes/fields can be added as needed.

While editing class.c for fun on an old VT220, it was obvious
that many lines still wrapped at 80 columns despite several
passes over the file in the past to correct it.  Finally
fixed the file, which involved minor rewrites in several
places to remove old code embedded within ifs and joined
together with &&.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>