JSR 292: optimise invokeBasic intrinsic
The method handle intrinsics are implemented by magicing
up native methods within resolveMethod.  The advantage of
this is that it required no changes to the interpreter
as they are invoked in the same way as a native method.
This speeded up initial development.  The other major
reason for doing it like this was that they also can be
reflected by a MemberName.

The problem is that to call the target they must reinvoke
the interpreter, which can lead to further calls to intrinsics.
With a very deep nesting, the C stack can overflow before
the Java stack has, leading to SEGV (several of the jtreg
tests deliberately provoke stack overflow).

The alternative is to implement the intrinsics as special
opcodes.  The intrinsics could then be implemented as a Java
method and invoked invisibly as now.  This would solve the
stack overflow issue and would also be faster.  However,
we would still have the overhead of creating a frame, etc.

The fastest way is to recognise the intrinsic methods, and
rewrite the initial invoke into the opcode for the intrinsic.
This is the approach taken here for the invokeBasic intrinsic
(only for the inlining interpreter so far).  We still magic
up a method as we still need to be able to reflect them.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>