JSR 292: internal opcodes for intrinsics
In addition to the invokeBasic intrinsic, the other
intrinsics now are implemented via internal opcodes
(linkToSpecial/Static and linkToVirtual).  This removes
the issue of recursion leading to the native C stack
being blown before the Java stack has overflowed.
The fast path is also ~20% faster at running the jtreg
tests.  The old intrinsics are still present - this
"slow" path will still be used when the intrinsics are
reflected.  This happens sufficiently rarely (at least
within the jtreg tests) that it's not worth slowing the
fast path down by checking for it.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>