JSR 292: implement intrinsic cache
Native method blocks are created for the various intrinsics
at different arities/return value size.  For the initial
implementation of JSR 292 in JamVM, new methods were created
for each call site, but this was wasteful of memory.

This change implements a simple cache using a hashtable.  A
new method is only needed if an equivalent does not already
exist in the cache (types do not need to match, only the
arity, return value size and intrinsic ID).  The entries
are reference-counted, and will be freed if they are unused
and the table becomes > 3/4 full.

A good example of the cache in operation is the jtreg test
MethodHandlesTest - previously 8799 methods were created.
With the cache this reduces to 271.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>