JSR 292: move support functions into own file
The JSR 292 implementation began by implementing the
native methods in MethodHandleNatives.  For this reason
the JSR 292 support ended up in natives.c.  However,
the support routines in terms of line count now far
exceed the relatively small number of native methods.
Of course, because of the pervasive nature of JSR 292,
code for it is scattered throughout the VM...

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>