JSR 292: Add LINKTOINTERFACE opcode
The initial implementation of the intrinsics used the same one for
linkToVirtual and linkToInterface.  The reason for this was that
there was an existing function within JamVM that did virtual
lookup (lookupVirtualMethod), and the initial implementation of
linkToVirtual just used the exising function.  However, the
existing function handles both virtual lookup and interface lookup,
so a separate implementation of linkToInterface wasn't needed.

But this is inefficient, as we end up doing additional checks
that aren't needed.  This change implements a separate opcode
for linkToInterface which enables linkToVirtual to be optimised
(linkToVirtual is the one we care about).

Note, intrinsics are implemented as both internal opcodes (fast
path) and as invocable methods (the slow path).  This change
does not bother to implement linkToInterface for the slow path
as this is not performance critical (it hardly gets used).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>