Allow classlib to have additional exceptions
Exceptions used within the VM are (mostly) loaded on
startup.  This change allows the list of exceptions to
be extended by the different class libraries.

Currently GNU Classpath adds no exceptions.  The reason
for the change is JSR292 on OpenJDK.  If JSR292 is enabled
BootstrapMethodError is added to the list.  This could
have been added to the "main" list if JSR292 was defined,
but the code requiring it is within the openjdk directory...

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>