JSR 292: tidy-up exports from reflect.c to mh.c
One of the things we need to be able to do is initialise a MemberName
from a reflection object (Constructor/Method/Field).  The problem
is the MemberName code is in mh.c, while the reflection internals
are within reflect.c (this includes references to the classes and
offsets of fields within them).  The reflection internals are only
used by functions within reflect.c and so are static.

Previously I just hacked them to be non-static, with the intention
of "doing it properly" with some accessor functions.  The problem
is reflect.c doesn't know about MemberNames and mh.c doesn't know
about reflection, so the code is inter-dependent and we're not
hiding much anyway.  So for now I've kept with allowing mh.c to
directly access the globals defined in reflect.c, but tidying it
up a bit.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>