JSR 292: implement MethodHandleNatives::expand
When the VM initialises or resolves MemberNames not all members are
filled in (as for most operations they are not required).  If these
members are required, the expand method will be called to construct
them lazily.  For the invoke jtreg tests this has not been
necessary, but for a full implemention expand must be implemented
(obviously, the jtreg tests in this respect are not complete).

For JamVM, the class member is always filled in, leaving the
name and type members for lazy computation.  This change expands
the name and type members as appropriate for methods/constructors
and fields.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>