GC compact: minor performance improvement + formatting
In threadChildren(), the class of the object is checked
to see if it is an instance of one of 3 "special" classes
that need to be handled specially.  This was done as
3 checks, which is done for every object.  This is changed
so that the checks are only done if the class is a special
class - if it is, we then determine how it is special.
This reduces the checks for the majority of objects to one
(but the special objects get an extra check - however,
as these are rare, overall the number of checks will be
substantially reduced).  This makes threadChildren()
match markChildren(), where this optimisation has already
been done.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>