JSR 292: Use int field for vmtarget on 32-bit machines
The vmtarget field is injected into MemberName by the VM.
This means we can use either an I or a J field depending on
the size of a pointer on the target.  This is an improvement
on VM fields "manually" added to classes, where the field
has to be big enough for any architecture (i.e. J).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>