JSR292: cleanup initialisation check in findMethodHandleType()
FindInvokeDynamicInvoker() can be called via an invokedynamic
bytecode before the MethodHandleNatives class has been
initialised.  This means initialiseMethodHandles will not have
been called as this is done via registerNatives, which is done
in MethodHandleNatives initialisation.

My previous hack simply called initialiseMethodHandles directly
in this case.  Normally this would have been OK, as in a
race condition, calling initialisation multiple times would be
benign.  However, initialiseMethodHandles also initialises the
intrinsic method hashtable.  In a race condition, if initialise-
MethodHandles were called twice we would get a memory leak.
However, in the worst case, however unlikely, there exists the
possibility of losing intrinsic methods already in the table.

This change replaces the direct call to initialiseMethodHandles
with a call to findSystemClass for MethodHandleNatives.  As
findSystemClass ensures the class is initialised we will get
initialiseMethodHandles called as side-effect.  However, the
class initialisation code correctly handles race conditions
and ensures that initialiseMethodHandles is only called once.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>