Re-work internal storage of annotation data
When I initially implemented annotation support in JamVM
I chose, in hindsight, a poor internal storage representation.
The extra overhead for annotations was a word in every
Class, Field and Method.  Minor attempts were made to minimise
the overhead in a method - each method can have 3 sets of
annotations - if any were present, these were stored in an
additional structure pointed to by the word within the method.
This meant the additional overhead per method, when no annotations
were present was 1 word, instead of 3.

The problem is most classes have no annotations at all.  These
end up with a NULL word in every method and field.  Overall, this
is a relatively large overhead for the common case (no annotations).

The new structure hangs everything off a single word in the class
structure.  Annotations on a method or field, if present, are
stored in a list (one list per annotation type), indexed by the
method or field number. This list is only allocated if at least
one annotation of the type is present.

This means there is no additional word in each method or field
structure.  If no method or field annotations exist, the overhead
is a single word per annotation type, in the annotation structure
pointed to by the word in the class.  If no annotations exist on
the class at all, the annotation structure is not allocated.
This means the entire overhead for a class with no annotations is
a single word in the class (which is NULL).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>