Implement JSR 308: Type Annotations
JSR 308 extends the existing Java annotation mechanism to
enable annotations to be made on any use of a type.  Two
additional annotation attributes are added to the class file -
RuntimeVisibleTypeAnnotations and RuntimeInvisibleTypeAnnotations.
Only the first needs to be recognised by the JVM.

As parsing of the annotation data is done in Java in OpenJDK, VM
changes are fairly trivial.  We just need to attach the data
to the appropriate class, method or type with as little overhead
as possible, and pass the data to Java as appropriate.

One slight issue is a seeming inconsistency in the Hotspot
implementation versus the JSR 308 spec.  The spec implies we can
have an annotation on a Code attribute in addition to a class,
method or field.  However, the Hotspot implementation of JSR 308
does not support this.  Browsing suggests that this follows the
output of javac, but that this may in fact be incorrect.  For
now I have followed the implementation in Hotspot.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>