JSR 292: make access.c conditional on JSR292 macro
Unlike hotspot, the same JamVM codebase supports OpenJDK
6, 7 and 8 (and GNU Classpath).  JSR 292 is only supported
on OpenJDK 8 (and future OpenJDK 7 when the backport appears).
Support for JSR 292 was made conditional on the JSR 292 macro.
However, I forgot to do this for access.c (I put it off because
the changes were messy, and then forgot to revisit it).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>