JSR 335 (lambda expressions): initial hack
JSR 335 (lambda expressions) requires VM changes to support
the new interface semantics (default methods, etc.).  Up to now,
I've got away with ignoring JSR 335, but it appears that the OpenJDK
class library has recently been updated to use these features.  For
example, with the latest version, half of the JSR 292 jtreg tests
now fail.

This initial checkin is a hack that enables the JSR 292 jtreg
tests to pass.  I believe I have worked out all the changes
needed for JSR 335, and I hope to implement them in the next day
or two.  However, I'm checking this initial hack in as a stop-gap
in case it takes longer...

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>