Yet another rename/refactor of annotation storage
I recently redid the annotation data structures to minimise
storage overhead in the case where annotations are not present.
I am very pleased with this.  However, it's come to my attention
that there's yet another new attribute that fits all the same
criteria as annotations (probably more so - it's unlikely to be
present as it's a new attribute in JDK8, and it must be enabled
on the javac command line).

So I can use my new super-duper annotation scheme.  The problem
is when I did it, it was only needed for annotations so all the
structures/macros are named to reflect this.  This change
renames/refactors things to make it more generic and reflect that
it can be used for any extra attribute.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>