JSR 308: make conditional on build-time JSR308 macro
I previously did not make the additional code for JSR308 conditional
at compile time.  Instead, if at runtime the type annotation fields
could not be found it didn't set them up.  This meant the code was still
present in the VM, but it would still work on class libraries without
JSR 308.

This is the way I used to handle things with GNU Classpath - for example,
having JamVM work with both the release version of GNU Classpath and the
HEAD.  This is because there was no way to configure things at build time
- it had to be done at runtime.

However, although JamVM's codebase has to work with different versions
of OpenJDK, it isn't the same binary - it is build separately for
each, and each one has its own configuration flag (openjdk6, 7, or 8).
This means we can limit the additional JSRs to the particular version
in which they were added.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>