JSR308/901: ifdef extra fields in ExtraAttributes
Both JSR308 and JSR901 add fields to the ExtraAttributes
structure of a class.  I didn't ifdef these initially
because the structure is a union with a void* array, which
enables the structure to be easily checked for non-null
fields (i.e. tell if it's needed or not).  At first, I
had this array defined as the same size as the number of
fields, so ifdefing the fields meant the size needed to change.
However, there's no reason at all to do this - the array
is simply a marker, and can be defined as size 0.  We work
out the number of fields by dividing the size of the union
by the pointer size.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>