JSR292: operand macros for direct-threaded interpreter
Macros must be defined to fetch the operands from the
new opcodes implemented for JSR292.  These must be defined
for each interpreter variant.  These were missing for the
direct-threaded interpreter (I only implemented them for
the code-copying JIT which is used by default, and the
indirect-threaded interpreter which I used for debugging).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>