Fix invokesuper check in invokespecial opcode
An invokespecial instruction is used to execute either
a nonvirtual call or to call a method in a superclass.
The rules to decide which is defined by the Java Virtual
Machine Specification (JVMS).  Surprisingly, the check used
by JamVM did not match the 1.7 version of the JVMS.  I have
checked back through earlier versions of JamVM, and this
code hasn't changed since at least JamVM 1.0.0 (March 2003).
So either it's always been wrong (which I hope is unlikely)
or the spec has changed in the meantime.  I will try and find
earlier versions of the JVMS to confirm.

It should be noted that this has never caused a problem (that
I'm aware of).  The difference came to light because of changes
in JSR335 (default interface methods).  With this, an
invokespecial can now be performed on an interface method.  As
an interface method it must be invoked non-virtually.  However,
in the case where the bug occurred, the interface method passed
JamVMs invokesuper check, and the method was invoked via the
superclass method-table (at best this will lead to a wrong
method being invoked, and potentially a SEGV).  With the check
from JVMS 1.7, this is handled correctly.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>