Do not free unprepared Miranda method code data
Unprepared Miranda methods share code data with the interface
method that they are a shallow copy of.  This means it cannot
be freed:

a) when/if an unprepared Miranda method is freed as a result
   of class GC

b) when the method is prepared and the now redundant bytecodes
   are freed.

Previously interface methods were all abstract, and their code
was set to a shared code segment that threw an abstract method
error if they were invoked.  As this is shared, abstract methods
did not free their code in the cases above.  As Miranda methods
are copies of interface methods, they were also abstract, and
therefore their code was not freed.

However, with JSR335, an interface method can now be a default
method.  In this case it is not abstract.  If a Miranda method
is a copy of this it will be freed incorrectly.  This means
we cannot free code data if the method is abstract, or is
a Miranda method.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>