JSR 292: tidy-up of intrinsic polymorphic method ID
The intrinsic methods previously had their ID stored in the
top 8 bits of the flags field.  Flags field has now been split
into a 1 byte state field, leaving 1 byte for flags (currently
only 2 bits used).  The polymorphic ID is now extra "states"
in the state field.

The reason for the change is to prepare for a tidy-up of method
preparation.  However, I've taken the opportunity to do some minor
clean-ups.  There now is an explicit ID_none, which is returned by
both polymorphicNameID and mbPolymorphicNameID - previously the
former returned -1, while the latter 0 (didn't actually matter,
as no checks relied on it).  Also, mbPolymorphicNameID is moved to
classlib-defs.h so it can be used in mh.c.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>