Split out common definitions between direct and inlining variants
The direct and inlining variants of the interpreter share
definitions for method preparation and accessing of
instruction operands.  These were duplicated in the
variant header files (interp-direct.h and interp-inlining.h).
A while ago I added new definitions for the JSR292 additional
opcodes to the inlining variant but forgot to add them to the
direct variant.  I added them to the direct variant in a
recent checkin, but these definitions should be in a shared
header file to stop this happening in the future.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>