JSR 292: polymorphic method ID (yet again)
Complete volte-face about polymorphic method IDs.  Not
the move from the flags field to the status field (which
is good) but the complexity (extra check) to return
ID_none - we don't need this as all checks are == or
relative.  This added an extra 40 bytes.  Normally this
would be OK, but it adds nothing (the extra code - albeit
tiny - meant it didn't "feel right").

Instead make IDs explicitly a state, and use the state field
"as is".

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>