Tidy-up of method preparation
As hinted in an earlier checkin, the reason for splitting
the flags field into a flags field and a new state field
is to tidy-up method preparation.

Previously, the "prepared" status of a method was stored
in the bottom bits of the code pointer itself.  The reason
for this was at the time there simply wasn't any field
in the MethodBlock structure that could be used.  The obvious
answer was to add one, but making the structure larger
was vetoed, when there was a "clever" alternative.

This worked fine, albeit a bit messy.  However, the "clever"
solution caused problems later when the inlining interpreter
was written.  Because of the way the bits were used, clear
bits could either be a prepared method, or a method in a
class that hadn't been linked.  Previously this didn't matter
but with the inlining interpreter we needed to know the
differance, so when freeing a method I had to an extra check
on the class status.

Much later extra room was needed for JSR 292, and it turned out
that it could be added without making MethodBlock larger
as there was a 2-byte "hole".  Later, a bug caused me to
notice the code needed to mask the bottom bits of the code
pointer, and I realised I no longer needed to do this as I
could use the extra room identified for JSR 292.

So instead of the bottom bits, we now have a one byte state
field, which explicitly stores the prepared status.  This
removes all the bit masking cruft, and the additional class
status check needed for the inlining interpreter.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>